style : directory name change : backend-kotlin -> backend
