--- backend-kotlin/src/main/kotlin/com/servy/user/UserApplication.kt
@@ -1,11 +1,0 @@
-package com.servy.user
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class UserApplication
-
-fun main(args: Array<String>) {
-	runApplication<UserApplication>(*args)
-}
