--- backend-kotlin/src/main/kotlin/com/servy/user/constants/ExceptionMessage.kt
@@ -1,10 +1,0 @@
-package com.servy.user.constants
-
-class ExceptionMessage {
-    companion object {
-        const val NOT_FOUND = "정보 조회하는데 실패하였습니다"
-        const val RUNTIME = "시스템 에러가 발생하였습니다"
-        const val INSERT_SUCCESS = "정보 저정하는데 성공하였습니다"
-    }
-
-}
