--- backend-kotlin/src/main/kotlin/com/servy/user/repository/ItemRepository.kt
@@ -1,7 +1,0 @@
-package com.servy.user.repository
-
-import com.servy.user.domain.Item
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface ItemRepository : JpaRepository<Item,Int> {
-}
