--- backend-kotlin/src/main/kotlin/com/servy/user/repository/OptionRepository.kt
@@ -1,7 +1,0 @@
-package com.servy.user.repository
-
-import com.servy.user.domain.Option
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface OptionRepository : JpaRepository<Option, Int> {
-}
