--- backend-kotlin/src/main/kotlin/com/servy/user/repository/QuestionRepository.kt
@@ -1,7 +1,0 @@
-package com.servy.user.repository
-
-import com.servy.user.domain.Question
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface QuestionRepository : JpaRepository<Question, Int> {
-}
