--- backend-kotlin/src/main/kotlin/com/servy/user/repository/SurveyRepository.kt
@@ -1,7 +1,0 @@
-package com.servy.user.repository
-
-import com.servy.user.domain.Survey
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface SurveyRepository : JpaRepository<Survey, Int> {
-}
