--- backend-kotlin/src/test/kotlin/com/servy/user/UserApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.servy.user
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class UserApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
