+++ backend/src/main/kotlin/com/survey/user/DTO/PageVo.kt
@@ -1,0 +1,10 @@
+package com.survey.user.DTO
+
+data class PageVo(
+    val pageNumber : Int,
+    val pageSize : Int,
+    val totalPages : Int,
+    val totalElements : Long,
+    val first : Boolean,
+    val last :Boolean,
+)
