+++ backend/src/main/kotlin/com/survey/user/UserApplication.kt
@@ -1,0 +1,11 @@
+package com.survey.user
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class UserApplication
+
+fun main(args: Array<String>) {
+	runApplication<UserApplication>(*args)
+}
