+++ backend/src/main/kotlin/com/survey/user/constants/ExceptionMessage.kt
@@ -1,0 +1,10 @@
+package com.survey.user.constants
+
+class ExceptionMessage {
+    companion object {
+        const val NOT_FOUND = "정보 조회하는데 실패하였습니다"
+        const val RUNTIME = "시스템 에러가 발생하였습니다"
+        const val INSERT_SUCCESS = "정보 저정하는데 성공하였습니다"
+    }
+
+}
