+++ backend/src/main/kotlin/com/survey/user/domain/Item.kt
@@ -1,0 +1,17 @@
+package com.survey.user.domain
+
+import jakarta.persistence.*
+
+@Entity
+@Table(name = "tb_item")
+class Item(
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Int = 0,
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "option_id", nullable = false)
+    var optionId: Option,
+
+    @Column(name = "item", nullable = true, length = 20)
+    var item: String,
+)
