+++ backend/src/main/kotlin/com/survey/user/domain/Survey.kt
@@ -1,0 +1,16 @@
+package com.survey.user.domain
+
+import jakarta.persistence.*
+
+@Entity
+@Table(name = "tb_survey")
+class Survey(
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Int = 0,
+
+    @Column(name = "title", length = 500, nullable = false)
+    val title: String,
+
+    @OneToMany(mappedBy = "surveyId", cascade = [(CascadeType.ALL)])
+    val questions: MutableList<Question> = mutableListOf(),
+)
