+++ backend/src/main/kotlin/com/survey/user/repository/ItemRepository.kt
@@ -1,0 +1,7 @@
+package com.survey.user.repository
+
+import com.survey.user.domain.Item
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ItemRepository : JpaRepository<Item,Int> {
+}
