+++ backend/src/main/kotlin/com/survey/user/repository/OptionRepository.kt
@@ -1,0 +1,7 @@
+package com.survey.user.repository
+
+import com.survey.user.domain.Option
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface OptionRepository : JpaRepository<Option, Int> {
+}
