+++ backend/src/main/kotlin/com/survey/user/repository/QuestionRepository.kt
@@ -1,0 +1,7 @@
+package com.survey.user.repository
+
+import com.survey.user.domain.Question
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface QuestionRepository : JpaRepository<Question, Int> {
+}
