+++ backend/src/main/kotlin/com/survey/user/repository/SurveyRepository.kt
@@ -1,0 +1,7 @@
+package com.survey.user.repository
+
+import com.survey.user.domain.Survey
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface SurveyRepository : JpaRepository<Survey, Int> {
+}
